move txn logic to hook
